Rename CFD-DEM tests that have restarts to start with "restart_" (#1400)
Description
While refactoring the DEM, @OGaboriault and I realized that the tests that have restart components are weirdly named. Some start with restart_, some end with _restart, and so on and so forth. This is a first step to uniformize the name of the tests that include restarts for CFD-DEM

Testing
Only the test names have been changed, nothing has changed regarding the results or the restart files themselves.