Resolves Rayleigh-Bénard Convection issue #1356  (#1403)
Description
The previous version of the Rayleigh-Benard example was a bit outdated. In this new version where we compare our simulation to results of the literature. We also produced another video for the simulation that is more fun to watch. We added a python script for the figures that follow the style of Lethe validation so it could be added to the tests if desired.

Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>
Co-authored-by: mivaia <59896845+mivaia@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>