Update cylinder examples in DEM with experimental results and fix rotating walls (#1422)
Description
The rotating drum example was not complete. We were not showing the comparison between the results of Lethe and the experimental results of Alizadeh. This was a regression from the original Lethe-DEM paper.
This PR fixes this by fixing the right friction coefficients and adding a full post-processing step which compares the velocity of the particles with the experimental results. This reverts the parameter values to the original Lethe-DEM paper and demonstrates that the agreement is still very good. The code has thus not regressed and I am a happy panda.

The issue mainly stemmed from the fact that the example did not have a rotating front and back wall. This was a major issue.

Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>