Dem epsd rolling friction model (#1417)
Description
A new rolling resistance model as been implemented. The elastic-plastic spring-dashpot (EPSD) rolling resistance model is similar to the hertz-mindlin models that we have in the sense that it keeps track of an overlap (or torque) over the duration of a contact in the contact_information container.

This model will be extremely useful when dealing with static DEM problems, as in the realization of an angle or a repose.

Testing
A new application test has been added for this new rolling resistance model. This test consists of three particles contacting each other with different initial angular velocity.

Documentation
New parameters as well as new theory has been added in the documentation.