Dem fix segfault solid surfaces (#1425)
Description
The moving_receptable_application test was segfaulting sometimes after #1417 on the CI. We think this is due to the clearing of the new tensor in the contact_info which should be cleared when there is no longer contact between particles.

The real cause of this segfault was the boost signal between the triangulation and the Insertion class that was still connected to the Insertion class when the triangulation was being destructed even though the Insertion class was already destructed prior.

Solution
Add the .clear() where it is needed. The new tensor had to be added to the IB_Particles contact_info for this to work.

A destructor was added to the Insertion class. This destructor disconnect the boost signal.