Prototype reinitialization vof (#1418)
Description
This PR adds a prototype for the geometric redistanciation. It includes an advection solver and the geometric redistanciation based on the article of Ausas et al. (2012), adapted to quads, parallel computation and adaptive mesh refinement. It also has a volume conservation algorithm. The treated advection field is a swirling flow, described in the same article.

Testing
Swirling flow case was mainly used to test the prototype. The position of the disk at the end of the simulation and its volume are the metrics of interest.
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>