Add interface reconstruction utilities (#1429)
Description
New functionalities for interface problems are added. The PR includes the computation of the volume enclosed by the level 0 of a level-set field and the surface reconstruction of the interface described by the level 0.

Testing
Two unit tests are added:

interface_tools_sphere_reconstruction: This test checks the interface reconstruction of the level 0 of a level-set field using the InterfaceTools::reconstruct_interface function. The level-set field of interest is the one describing a sphere. The reconstruction is computed for 3 mesh refinements and the test checks the error on the area of the reconstructed surface triangulation and the convergence rate of the method (formally 2). The area is computed using dealii GridTools::volume() method.

interface_tools_sphere_volume: This test checks the computation of the volume enclosed by the level 0 of a level-set field using the InterfaceTools::compute_volume function. The level-set field of interest is the one describing a sphere. The volume is computed for 3 mesh refinements and the test checks the error on the volume and the convergence rate of the method (formally 2).

Documentation
All the briefs are added