Remove muparser (#1452)
Description
The custom parsed function and muParser dependency are removed. They had been added in a previous PR, but it has been unused for some time so it is being removed.

Testing
The associated unit test is deleted. All other tests still pass.

Documentation
The associated documentation was adapted to take into account the removal.