Matrix-free implementation of the Volume-Averaged Navier-Stokes equation form A (#1448)
Description
This PR adds a matrix-free implementation of the volume-averaged Navier-Stokes equation. This PR only implements form A of these equation, since this is the one we use the most. Adding form B of the equation in the future will be easy to achieve.

This PR essentially adds four source/header files:
fluid_dynamics_vans_matrix_free.cc/h
fluid_dynamics_vans_matrix_free._operators.cc/h

The file contains a very barebone matrix-free solver for the VANS equation. The code is very light because the class inherits from the regular matrix-free solver which removes a lot of code duplication. This was relatively easy to implement.

The second file adds the matrix-free operator. This is the VANS matrix-free operator with grad-div stabilization for form A of the VANS equation. Right now the only thing that is supported is that the void fraction is established using a function instead of being established from particles. Establishing it with particles is possible with the void fraction manager, but more work will be required to actually add this because it means that the MatrixFree object needs to point to two different dof_handler. AKA this needs some major refactoring that will be done in future PRs.

Testing
This is tested like crazy. I have added 7 application_tests for the various cases of MMS we use as well as a very sharp tanh void fraction jump to test if the solver works well. The solver does work very well.

Documentation
This feature is very experimental, I would rather not document it yet because it is not ready to be used at all.



* Add missing operator file

* First implementation with constant void fraction

* WIP checkpoint before grad terms

* Velocity converging not pressure

* Convergence is good, some error in jacobian matrix still

* Finished working version

* Add new test cases

* Fixed an error in the RHS

* Update test results

* Add CFD-DEM parameters

* Add grad-div stabilization

* Add sanity checking for VANS model

* Document matrix free operator

* Finalize documentation and indent

* Fix initialized variables

* Fix documentation of the weak form

* Fix missing *

* Update changelog

* Fix a nasty bug with evaluator update

* add missing tests

* Update failed test results

* Add missing test

* Indent all the PRM files

* Apply suggestions from code review

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>

* Fix mapping

* Reput mapping and reduce length of two tests

* Apply suggestions from code review

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>

* Remove unused function

* Add more verbose comment

* Apply suggestions from code review

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>

* Add better comments for the VANS operator

* remove useless parameters from tests and re-indent

* Increase timeout for tanh case

* Apply suggestions from code review

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>

* Indent

* Update include/fem-dem/fluid_dynamics_vans_matrix_free.h

Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>

* Update source/fem-dem/fluid_dynamics_vans_matrix_free_operators.cc

Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>

* Refactor the concatenate string into utilities

---------

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>