Vof geometric reinitialization architecture (#1451)
Description
This PR adds the (almost) empty architecture of the SignedDistanceSolver and its coupling in the VOF solver (geometric reinitialization method ) , following the structure of the already existing vof_advection prototype. The parameters associated to the geometric reinitialization method for the VOF solver are also added.

Some methods of the SignedDistanceSolver are implemented to be able to test this new architecture.

Testing
One unit is implemented:

interface_tools_signed_distance: This test checks the instanciation of the InterfaceTools::SignedDistanceSolver class from within an abitrary structure that replicates the usual physical solver structure. This is a dummy test intended to check the architecture, and not the SignedDistanceSolver performances yet! When the implementation of the SignedDistanceSolver will be completed, this test will be extended to a verification of the signed distance computation for a sphere.

Documentation
The new parameters are not documented yet because they are not currently used. They will be in the next PR implementing the solver itself.

* Add iso-level class member and exception

* Update changelog

* Fix test with dealii vectors

* Fix test with dealii vectors completely

* Indent

* Apply comments from code review

* Fix bug introduced by remove mpi_comm arg