Core library cleanup (#1457)
Description
In the aims of deploying a CI that would check for warnings using the clang compiler, I am fixing some of these warnings ahead of time by cleaning the code file by file. This PR addresses this for only 4 files in the core library, but this has impact beyond the code library itself.

One important thing, I have learnt that according to clang we should not mark stuff const in the .h file, but on in the .cc file ( see https://stackoverflow.com/questions/38660537/why-is-const-unnecessary-in-function-declarations-in-header-files-for-parameters) and so I have applied these changes also here.