Example unresolved cfd dem sedimentation (#1450)
Description
This example simulates the sedimentation of a single particle in a water tank due to gravity. As with other unresolved CFD-DEM applications and examples, we start it but introducing the particle in our domain by running a DEM simulation (using lethe-particles). The DEM simulation generates checkpoint files that will be used by our unresolved CFD-DEM application (lethe-fluid-particles) to determine the initial position of the particle.

This PR also fixes a bug on the particle insertion method. Before, if we had a single time-step simulation for the particle insertion with a insertion frequency of 1, the particles wouldn't be inserted. Now this exception is handled.

Lastly, a test was added to make sure this behavior is not modified for this exception.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>