Example 2D Sudden Expansion Flow (#1455)
Description
This example simulates an incompressible flow through a two-dimensional channel with a sudden expansion. A low Reynolds number (Re=70) is employed for a steady-state simulation, and a high-Reynolds number (Re=610) is used for observing transient results. In both situations, previous numerical results are compared to the output herein obtained.