Throw exception for non-Newtonian fluid in MF solver (#1463)
Description
The matrix-free solver does not support non-Newtonian fluids at the moment. However, if such fluid parameters were included in the prm file, the simulation would still run.

Solution
An exception verification was added to the MF solver so that, if non-Newtonian fluid parameters are passed, the simulation is interrupted and an exception message is shown.

Testing
Previous example files using the lethe-fluid-matrix-free solver were tested with entries for non-Newtonian fluid before and after the fix. Before the fix the files would still run, while after the fix an exception is shown and the simulation is aborted.