Fix kokkos compilation for new syntax changes (#1470)
Description
The deal.II PR 18273 (dealii/dealii#18273) has changed the syntax of the Kokkos interface. This prevents compilation of our kokkos prototypes

Solution
This PR fixes this by adapting the syntax of the examples to the new syntax. This is a minor change and the examples still work just fine.