Add feature for last output and fix example results (#1471)
Description
In transient simulation, the last result of a transient simulation would not be written as VTU file if it did not correctly fall onto the output frequency.

Solution
This PR fixes this by always writing the last VTU output when we are doing a transient simulation with a frequency of output.
I used to opportunity to also fix the documentation for the rising bubble and add the results we are obtaining with Lethe with this new feature. Now the comparison of the bubble with the reference result is just absolutely perfect meaning we are good.

Testing
No tests should be changed