Thermal DEM temperature integration (#1473)
Description
This PR is to add an integration function for temperature for thermal DEM and a unit test to make sure it works.

Testing
A new test integration_temperature.cc is added to check the order of the temperature integration (should be one for Euler method), by solving mcdT/dt = -T for two time steps.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>