Apply filter before starting reinitialization process (#1474)
Description
Phase fraction gradient projection is computed with the filtered phase fraction gradient. However, before the starting the VOF algebraic interface reinitialization process, the filter was not applied leading to wrong values of filtered phase fraction gradient and therefore curvature. This is now fixed.

Also, in the calculate_momentum function of VOF post-processing quantities, the FEValues of fluid dynamics was initialized with the FE degree of the fluid dynamics physics instead of the VOF one leading to a vector size error when getting the function values. The quadrature formula has now been changed to the VOF one.

Solution
Applying the VOF filter to the phase fraction field before computing the phase fraction gradient projection and before starting the reinitialization process.

Testing
The fix was tested with two examples: rising bubble and Rayleigh-Taylor. For both cases, no visible differences are observed.
A new application test with an advected circle was added. The test uses the algebraic interface reinitialization feature and Q1Q2Q1 elements (phase fraction, velocity, pressure).



* Correct typo

* Fix quadrature formula inconsistency