Add options to print version and parameters (#1475)
Description
This PR adds two new features:

Allows to print the deal.II and Lethe versions used to run an application. An example of the usage would be as follows:  lethe-fluid -V yields:
Running: lethe-fluid -V
lethe/1.0-77-g9fe541d99 deal.II/9.6.2-2966-g7a74e69540
Kudos to @paulapatience who suggested this format and contributed to the implementation in deal.II.

Allows to print the parameters of the input file by specifying an additional parameter as follows:
#---------------------------------------------
# Print non-default parameters to output file
#---------------------------------------------

set print parameters = only changed
Options are none (default), all, or only changed (only supported with deal.II versions > 9.7).

Testing
The features were tested for the examples of all the applications.

Documentation
I have not introduced this in the documentation yet as there is no section for such parameters. We can discuss whether such a section is really needed.