Fix refinement frequency bug in Sharp IB solver (#1479)
Description
Before this PR, lethe-fluid-sharp would attempt to refine the cells close to the immersed boundary at every time step. This behavior is desired if solids moved since the last time step, since it ensures that the interface is properly captured regardless of the regular mesh adaptation frequency.

When solids are stationary, this behavior becomes costly if the mesh is large or if solids are numerous or complex.

Solution
An additional condition is checked at the beginning of the function that refines immersed boundaries.

Testing
All tests still pass.

Costly simulations on computing clusters now run at a lower cost.