Read and merge meshes in mortar control (#1481)
Description
A mortar control subsection has been added to the code, based on the features from pull request #1462.
The goal of using the mortar control is to simulate the interaction between two domains, i.e. a rotor and a stator part. In this sense, this PR introduces the mortar parameters to be read in the prm file, as well as defines a function able to read and merge two meshes. Currently this feature is accessible through the lethe-fluid solver, and is restricted to dealii mesh types only. This will be adjusted in future implementations.

Testing
Two new tests were added considering different geometries. The goal is solely to test the function read_mesh_and_manifolds when the mortar control is enabled.

Documentation
In the new subsection Mortar Control, the added parameters are:

Secondary mesh: while the standard mesh read in the prm file is called the stator (outer) domain, a new (rotor) mesh is now read;
Boundary IDs of the rotor/stator matching interfaces;
Center of rotation for the rotor domain.


* Update parameter description

* WIP test for mortar meshes

* WIP adapt solver to read mortar data

* WIP attach triangulation to mortar meshes

* Move test for reading mortar data to tests/mortar

* Update test

* Update test output

* Attach manifold only at merged triangulation

* Fix default mortar parameter

* Delete unecessary functions

* Add description on read_mesh_and_manifolds

* Add exception for gmsh case

* Rotate rotor grid

* Remove (temporarily) setup for periodic BCs

* Update test description

* Update test output file

* Fix input parameters

* Use read mesh function in test

* Add test for new geometry

* Rename test

* Add initial rotation angle in mortar rst file

* indent all

* Indent code section in rst file

* Delete unused parameters

* indent all

* Update test output file

* Automate setting manifolds in merged triangulation

* indent all

* Rename read mesh and manifolds function for mortar case

* Move variables descriptions inside the function

* Change tolerance on merging triangulations

* Update test output file

* Change variables to const

* indent all

* Update mortar section name

* Update descriptions

* indent all