GMG for matrix-free VANS (#1458)
Description
The previous implementation of the multigrid preconditioner for the VANS equation was not assembling the correct operator on each level, this was problematic (and it was actually kind of a miracle that it worked).
This PR fixes this by evaluating the operator on every grid level correctly.

Testing
All the tests still pass, but the number of iteration makes a lot more sense now.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>