Particle particle heat transfer (#1478)
Description
This PR adds the functions that will be used for thermal DEM including the calculation of the thermal resistances, the total conductance and two functions to apply heat transfer to particles for local contacts or ghost-local contacts, in the same way it is done for the contact force.

Testing
A test of the calculation of the thermal resistances is added.

Documentation
Miscellaneous (will be removed when merged)
For now, these functions can only be used directly in tests.
I plan on using them in the particle_particle_contact_force files next.

* split into .cc and .h

* correct

* corect error

* add const

* modify erfc-1 function

* add comments and change erfc-1 function

* adjust test output

* apply comments

* replace sqrt