Fix the plot in the rotating drum and refactor (#1485)
Description
The validation script plot for the rotating drum had a bit of a weird error in the legend (see picture)
image

Furthermore, the validation scripts contained a lot of duplicated functions.

I have fixed the legend and refactored the duplicated function to be a in a seperate file. This makes deploying a new validation much faster.

Testing
Relaunched validation