Mortar: test oversampling (#1488)
Oversampling (in our case) means that you take more quadrature points for the coupling than for other integrals. This is for example done in preCICE (a coupling library). See: https://link.springer.com/chapter/10.1007/978-3-030-47956-5_10, where they used a factor of 3. Not sure if we will need this, but being able to easily test this does not harm.