Fix level set output when geometry is moving (#1495)
Description
The caching mechanism for the level set calculation prevented the recalculation of the combined_shape level set when the particles were moving.

Solution
Added a check that forces clearing the level set cache of the combined_shape for the cases where particle motion is imposed OR particle integration.

Testing
No test results should change. I'm not sure how to test this adequately at the moment so that this bug never happens again, I'll try to figure out a test in a follow-up PR. This was a visualization bug.


Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>