Fix rayleigh benard example (#1496)
Description
The Rayleigh-Bénard example would not yield the correct profile on my machine at home for the Ra=10k case and would instead lead to a quasi-static velocity field. This PR ensures that an instability is always generated by setting a linear initial condition. Furthermore, using this approach, the steady_bdf time-stepping scheme can be used to avoid going to very fine meshes and/or long running time. Using this, the simulation is much faster and the stopping criteria is independent of the end time. This actually saves a lot of time!