Mortar update merged triangulation data (#1490)
Description
The mortar subsection previously added in pull request #1462 is revised in this new PR. the main change is the storage of manifolds in the rotor/stator and merged triangulations, which has been fixed to account for all faces and cells. Also, a sanity check has been added to verify if the number of faces at the rotor and stator matching interfaces is the same.

Testing
The two tests previously added for checking the read_mesh_and_manifolds_for_stator_and_rotor still pass, as well as all other tests.

Documentation
The documentation has been updated to better explain the identification of the rotor/stator matching interfaces.