Template particle insertion and add initial temperature (#1491)
Description
This PR adds multiphysic DEM properties, like temperature, to particle insertion. PropertiesIndex is added as a template parameter of Insertion classes.
For volume and plane insertion, the initial temperature is a parsed function of position.

Testing
A test insertion_initial_temperature is added to check that the initialization of the temperature with a function works well.

Documentation
Documentation for insertion in multiphysic DEM will be added later.


Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>