Add VOF geometric reinitialization solver (#1497)
Description
This PR completes the signed distance solver by filling the (almost) empty architecture of the SignedDistanceSolver implemented in PR #1451. Its coupling in the VOF solver (geometric reinitialization method) is also completed. The implemented routines follow the one of the already existing vof_advection prototype.

Testing
A new unit test is added to check the signed distance solver. It checks the InterfaceTools::SignedDistanceSolver class from within an arbitrary structure that replicates the usual physical solver structure, denoted by the background prefix. This test performs verification of the signed distance computations for a sphere.

A new application test is added. It tests the swirling-flow case using the VOF solver with geometric redistanciation.

Previous application tests outputting the VOF mass conservation have been updated since the geometric volume is added to the output table.

Documentation
The documentation will be updated in a future PR.
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>