Add find-line-sphere-intersection (#1511)
Description
This PR Adds the find_line_sphere_intersection in lethe_grid_tools for use in the case of ray-particle intersection for the profilometry hackathon project. This function simply evaluates all intersection points between a line and a sphere.

Testing
The methodology was tested locally for both 2D and 3D intersection.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>