Update mms case (#1506)
Description
The MMS example is very valuable because it clearly demonstrates the convergence of the code. However, making the plot for the final results relied on a python script that had a lot of copy-pasted elements. Furthermore, the python script to launch the simulations struggled on my machine and kept relaunching some simulations that were completed.
I addressed these two points through some refactoring.

Refactored the plotting script to remove as much code duplication as possible. This was achieved by adding some functions that are reused. The final code is significantly smaller. I also made it such that the code outputs the plots at the screen also so a user can readily see the results.

I added .sh launching script to ensure that the simulations can launched in a more portable fashion than by using a python script.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>