Add multiphysic DEM solver type (#1514)
escription
Multiphysic DEM simulations can now be launched with solver type dem_mp to follow the evolution of the temperature of the particles.
Temperature is also added in the properties printed by application tests for solver type dem_mp.

Testing
An application test is added, it is the same as the packing_in_ball test but it uses multiphysic DEM.

Documentation
The documentation for multiphysic DEM will be added in a future PR.