Discontinuous Galerkin implementation for VOF (#1510)
Description
This PR adds discontinuous Galerkin framework to the VOF solver. Thus far, the framework is compatible with surface tension and with pretty much everything except the redistanciation of the solvers (for which it has not been tested at all).

The main issue with the solver is that the results thus far are very underwhelming. It does not collaborate very well with the Kelly error estimator AND it does not collaborate well with the BDF2 time integrator. My goal with this is so finally set the basis for the solver cleanly, but this will remain an active area of research in the next few months.

Testing
I added a simple advection test case with the DG solver that is a very good measure of the accuracy. This will be enough to make sure the solver is maintainable in the short term.

Documentation
Since this feature is very experimental, I would rather not document it and expose it to the users until it leads to good results. If I cannot get good stuff with it this year, I would maybe remove it.


Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>