Add buoyancy term to matrix-free operators and GMG (#1524)
Description
This PR adds the implementation of the buoyancy term to the matrix-free operators, by implementing the following functionalities:

Add a function that gets the temperature solution from the multiphysics interface and converts it to deal.II vectors.
Precompute the buoyancy term using an FEValues object and a temperature DoFHandler. This is precomputed for its usage within the cell and residual loop of the operator.
In order to use the GMG preconditioner an additional function initialize_auxiliary_physics(...) has been created. It transfers the current temperature solution to the different MG levels.
Testing
No changes in current tests should be observed. A test should be added in the future since more testing is needed regarding the effect of the buoyancy term into the smoothers/coarse-grid solution of the GMG.

Documentation
Nothing has been added to the documentation yet, since it is still a feature that needs to be tested.