Remove template argument from CouplingOperatorBase (#1527)
Remove DataType template argument from CouplingOperatorBase. CouplingOperatorBase now only works on sets of doubles. The doubles need to be converted to/from the actual data types during evaluation/integration, e.g., with the help of BufferRW.

This PR is needed to be able to move the implementation to a source file.