Non-Newtonian Matrix-Free Operator (#1523)
Description
This PR couples the implementation of non-Newtonian models with the matrix-free solver.
Instead of accessing solely the kinematic viscosity, the NavierStokesOperatorBase now access the physical properties manager, which allows the update of viscosity according to the rheological model.

Testing
Two new tests have been added, using the Method of Manufactured Solutions with FE approximations of first and second order. The tests were adapted from the lethe-fluid application tests, with the only difference that now the multigrid preconditioner is chosen.

Documentation
No parameters were introduced.



Co-authored-by: voferreira <voferreira@outlook.com>
Co-authored-by: Laura Prieto Saavedra <lauraprietosaavedra@gmail.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>