Refactor particle-wall contacts (#1520)
Description
This PR changes the architecture of the code for particle-wall contacts to reproduce the one for particle-particle contacts and remove code duplication. The force calculation is templated in the same way as done for particle-particle contacts, i.e. with the contact model and rolling resistance model.
The particle_wall_contact_info has been changed to a struct instead of a class and normal_overlap, tangential_relative_velocity and normal_relative_velocity_value are calculated on the fly and not stored in contact_info anymore.
This should close #1233 issue.
The calculation of force_on_walls and torque_on_walls has been removed for now.

Testing
Unit tests of particle-wall contact forces were modified (and cleaned up) and all pass. Application tests also pass.